third_party/blink/web_tests/fast/dom/shadow/remove-and-insert-style-expected.txt

Test for Issue: 246300: Styles in nested shadows are not recalculated correctly on insertion.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS backgroundColor is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
